+++ resources/ident.rb
@@ -16,6 +16,7 @@
 # limitations under the License.
 #
 provides :postgresql_ident
+unified_mode true
 
 property :mapname,     String, required: true
 property :source,      String, default: 'pg_ident.conf.erb'