+++ resources/user.rb
@@ -15,6 +15,7 @@
 # limitations under the License.
 #
 provides :postgresql_user
+unified_mode true
 
 property :create_user,        String, name_property: true
 property :superuser,          [true, false], default: false